Add files via upload
These codes were developed by Alex Ritzie, an undergraduate physics student in the Fall 2023 class. Thanks to Alex forr sharing these codes. It identifies the outliers, and calculates the average and standard deviation in a data set. Data has to be entered as a one-dimensional array. Future work will include adding the ability to read data from a csv or xlsx file. Watch for updated versions.